% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_exported.R
\name{get_parscale}
\alias{get_parscale}
\title{Get parameter scales for optimization}
\usage{
get_parscale(start)
}
\arguments{
\item{start}{numeric vector of start values of coefficients}
}
\value{
numeric vector of scales, each being a power of 10
}
\description{
Results are useful as an argument to the \code{parscale} component of \code{control} argument of \code{optim} function (and other optimization functions).
}
\examples{
\dontrun{
get_parscale(c("(Intercept)" = 3.8, "educ" = 0.142, "exper" = 0.027, "tenure" = 0.021))
}
}
