% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_exported.R
\name{named_list}
\alias{named_list}
\title{Create a list whose elements have the same names as the variables in input.}
\usage{
named_list(...)
}
\arguments{
\item{...}{objects}
}
\value{
a named list containing given objects
}
\description{
Create a list whose elements have the same names as the variables in input.
}
\examples{
\dontrun{
x <- 10
y <- "letter y"
named_list(x, y) # same result as list(x = x, y = y)
}
}
