% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ttsfa}
\alias{ttsfa}
\title{Two-Tier Stochastic Frontier Analysis}
\usage{
ttsfa(
  formula,
  data,
  start_val = NULL,
  ineff_distr = c("exp", "exp"),
  error_param_min = 0.01,
  error_param_max = 0.5,
  maxit = 1e+05,
  trace = 0L
)
}
\arguments{
\item{formula}{a symbolic description of the model to be estimated}

\item{data}{data.frame that contains cross-section data}

\item{start_val}{numerical vector of initial values for optimizer; if NULL (default), will use OLS estimates}

\item{ineff_distr}{2-element character vector specifying the distributions of the inefficiency (asymmetric) error terms; currently only \code{c("exp", "exp")} is supported}

\item{trace}{integer in range 0:6; higher values cause more information on the progress of the optimization to be printed}
}
\value{
\describe{
\item{mle_fit}{an object of class \code{mle2}; see \code{\link[bbmle]{mle2}}}
\item{ineff_distr}{equal to argument \code{ineff_distr}}
\item{unexplained_variation}{numeric value of unexplained variation in dependent variable, in range \link{0, 1}}
\item{unexplained_variation_from_inefficiency}{numeric value of ratio of unexplained variation that comes from inefficiency (asymmetric error terms), in range \link{0, 1}}
\item{surplus}{list containing observation-specific surplus estimates from the composed error term:
\describe{
\item{u}{negative extracted surplus = E(u|epsilon)}
\item{w}{positive extracted surplus = E(w|epsilon)}
\item{exp_neg_u}{percentage measure of price reduction = E(exp(-u)|epsilon), if log(price) is dependent variable}
\item{exp_neg_w}{percentage measure of price increase = E(exp(-w)|epsilon), if log(price) is dependent variable}
}
}
\item{ols_fit}{an object of class \code{lm}; see \code{\link[stats]{lm}}}
\item{formula}{equal to argument \code{formula}}
}
}
\description{
Two-Tier Stochastic Frontier Analysis
}
\examples{
\dontrun{
# Replication of Kumbhakar, S. C., & Parmeter, C. F. (2009), <doi:10.1007/s11123-008-0117-3>
library(wooldridge) # for dataset wage2
data(wage2)
d <- na.omit(wage2)
d$educ_sq <- d$educ^2
d$exper_sq <- d$exper^2
d$tenure_sq <- d$tenure^2

fit <- ttsfa(formula = lwage ~ educ + educ_sq + exper + exper_sq + tenure + tenure_sq +
               age + IQ + urban + married + south + sibs + brthord + meduc + feduc,
             data = d)
summary(fit)
}
}
